renaming related fragment files
