+++ java/app/src/main/java/com/fullstorydev/shoppedemo/ui/cart/CartFragment.java
@@ -1,4 +1,4 @@
-package com.fullstorydev.shoppedemo.ui.home;
+package com.fullstorydev.shoppedemo.ui.cart;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -12,17 +12,17 @@ import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 import com.fullstorydev.shoppedemo.R;
 
-public class HomeFragment extends Fragment {
+public class CartFragment extends Fragment {
 
-    private HomeViewModel homeViewModel;
+    private CartViewModel cartViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
             ViewGroup container, Bundle savedInstanceState) {
-        homeViewModel =
-                ViewModelProviders.of(this).get(HomeViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_home, container, false);
-        final TextView textView = root.findViewById(R.id.text_home);
-        homeViewModel.getText().observe(this, new Observer<String>() {
+        cartViewModel =
+                ViewModelProviders.of(this).get(CartViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_cart, container, false);
+        final TextView textView = root.findViewById(R.id.text_cart);
+        cartViewModel.getText().observe(this, new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
