+++ java/app/src/main/java/com/fullstorydev/shoppedemo/ui/market/MarketViewModel.java
@@ -1,16 +1,16 @@
-package com.fullstorydev.shoppedemo.ui.home;
+package com.fullstorydev.shoppedemo.ui.market;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class MarketViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public MarketViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is market fragment");
     }
 
     public LiveData<String> getText() {
