+++ java/app/src/main/res/layout/fragment_market.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent" >
  
     <TextView
-        android:id="@+id/text_dashboard"
+        android:id="@+id/text_market"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
