+++ java/app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
     <string name="app_name">ShoppeDemo</string>
-    <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
+    <string name="title_market">Market</string>
+    <string name="title_cart">Cart</string>
 </resources>
